requirements.txt
requests
dotenv
psycopg2-binary